test: extract generator tests into their own files
