+++ packages/markdown-it-image-size/test/index.test.ts
@@ -37,44 +37,6 @@ describe(markdownItImageSize.name, () => {
     expect(actual).toBe(expected);
   });
 
-  it("should use 11ty inputPath to resolve relative paths if available", () => {
-    const inputPath = "./test/test-assets/posts/1/1.md";
-
-    const imageUrl = "./post-image.jpg";
-    const markdown = `![](${imageUrl})`;
-
-    const imageWidth = 4032;
-    const imageHeight = 3024;
-
-    const expected = `<p><img src="${imageUrl}" alt="" width="${imageWidth}" height="${imageHeight}"></p>\n`;
-    const actual = markdownRenderer.render(markdown, {
-      page: {
-        inputPath,
-      },
-    });
-
-    expect(actual).toBe(expected);
-  });
-
-  it("should use VitePress path to resolve relative paths if available", () => {
-    const markdownRenderer = new MarkdownIt().use(markdownItImageSize);
-
-    const path = "./test/test-assets/posts/1/1.md";
-
-    const imageUrl = "./post-image.jpg";
-    const markdown = `![](${imageUrl})`;
-
-    const imageWidth = 4032;
-    const imageHeight = 3024;
-
-    const expected = `<p><img src="${imageUrl}" alt="" width="${imageWidth}" height="${imageHeight}"></p>\n`;
-    const actual = markdownRenderer.render(markdown, {
-      path,
-    });
-
-    expect(actual).toBe(expected);
-  });
-
   it("should render external images with attributes for width and height", () => {
     const imageUrl =
       "https://images.unsplash.com/photo-1577811037855-935237616bac?auto=format&fit=crop&w=2167&q=80";
