+++ packages/markdown-it-image-size/test/option-public-dir.test.ts
@@ -3,7 +3,7 @@ import { beforeEach, describe, expect, it } from "vitest";
 import { markdownItImageSize } from "../src";
 import { clearCache } from "./test-utils";
 
-const cacheFile = "cache-5.json";
+const cacheFile = "option-public-dir-test.json";
 
 describe("option publicDir", () => {
   beforeEach(() => {
